rename sensor class, give type_ attribute and move sensor classes and methods to sensors.py
