+++ tests/instruments/test_sensors.py
@@ -5,7 +5,7 @@ from virtualship.instruments.sensors import (
     SensorType,
 )
 from virtualship.instruments.types import InstrumentType
-from virtualship.models.expedition import SensorConfig
+from virtualship.models.expedition import SENSOR_REGISTRY, SensorConfig
 from virtualship.utils import get_supported_sensors
 
 EXPECTED_SENSOR_MEMBERS = {
@@ -20,6 +20,36 @@ EXPECTED_SENSOR_MEMBERS = {
     "PHYTOPLANKTON",
     "PRIMARY_PRODUCTION",
 }
+
+
+def test_sensor_registry_keys_match_sensor_type():
+    """SENSOR_REGISTRY keys must be exactly the set of SensorType members."""
+    assert set(SENSOR_REGISTRY().keys()) == set(SensorType)
+
+
+@pytest.mark.parametrize(
+    "sensor_type",
+    [
+        SensorType.OXYGEN,
+        SensorType.CHLOROPHYLL,
+        SensorType.NITRATE,
+        SensorType.PHOSPHATE,
+        SensorType.PH,
+        SensorType.PHYTOPLANKTON,
+        SensorType.PRIMARY_PRODUCTION,
+    ],
+)
+def test_sensor_registry_bgc_entries_category(sensor_type):
+    """All BGC sensors must have category 'bgc'."""
+    assert SENSOR_REGISTRY()[sensor_type].category == "bgc"
+
+
+def test_sensor_registry_unique_fs_keys():
+    """No two sensors should share an fs_key."""
+    fs_keys = [meta.fs_key for meta in SENSOR_REGISTRY().values()]
+    assert len(fs_keys) == len(set(fs_keys)), (
+        "Duplicate fs_key found in SENSOR_REGISTRY"
+    )
 
 
 def test_sensor_type_all_members_exist():
