+++ tests/test_utils.py
@@ -4,16 +4,15 @@ from pathlib import Path
 import numpy as np
 import pytest
 import xarray as xr
-from parcels import FieldSet, JITParticle, ScipyParticle, Variable
 
 import virtualship.utils
+from parcels import FieldSet, JITParticle, ScipyParticle, Variable
 from virtualship.instruments.sensors import SensorType
 from virtualship.instruments.types import InstrumentType
 from virtualship.models.expedition import Expedition, SensorConfig
 from virtualship.models.location import Location
 from virtualship.utils import (
     PROJECTION,
-    SENSOR_REGISTRY,
     _calc_sail_time,
     _calc_wp_stationkeeping_time,
     _find_nc_file_with_variable,
@@ -363,36 +362,6 @@ def test_calc_wp_stationkeeping_time_no_instruments(expedition):
 
     assert stationkeeping_null == stationkeeping_emptylist  # are equivalent
     assert stationkeeping_null == datetime.timedelta(0)  # at least one is 0 time
-
-
-def test_sensor_registry_keys_match_sensor_type():
-    """SENSOR_REGISTRY keys must be exactly the set of SensorType members."""
-    assert set(SENSOR_REGISTRY().keys()) == set(SensorType)
-
-
-@pytest.mark.parametrize(
-    "sensor_type",
-    [
-        SensorType.OXYGEN,
-        SensorType.CHLOROPHYLL,
-        SensorType.NITRATE,
-        SensorType.PHOSPHATE,
-        SensorType.PH,
-        SensorType.PHYTOPLANKTON,
-        SensorType.PRIMARY_PRODUCTION,
-    ],
-)
-def test_sensor_registry_bgc_entries_category(sensor_type):
-    """All BGC sensors must have category 'bgc'."""
-    assert SENSOR_REGISTRY()[sensor_type].category == "bgc"
-
-
-def test_sensor_registry_unique_fs_keys():
-    """No two sensors should share an fs_key."""
-    fs_keys = [meta.fs_key for meta in SENSOR_REGISTRY().values()]
-    assert len(fs_keys) == len(set(fs_keys)), (
-        "Duplicate fs_key found in SENSOR_REGISTRY"
-    )
 
 
 # helper
